"Take" works and now you can create programs that never stop running.
