+++ app/actors/messages/ActorFailed.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class ActorFailed(msg: String = "Json did not match any type.")
