+++ app/actors/messages/CompilerCreate.scala
@@ -1,0 +1,5 @@
+package actors.messages
+
+import model.models.Problem
+
+case class CompilerCreate(steps : Int, problem : Problem)
