+++ app/actors/messages/CompilerExecute.scala
@@ -1,0 +1,6 @@
+package actors.messages
+
+import model.models.Problem
+
+case class CompilerExecute(steps : Int, problem : Problem)
+
