+++ app/actors/messages/CompilerHalt.scala
@@ -1,0 +1,7 @@
+package actors.messages
+
+class CompilerHalt()
+
+object CompilerHalt {
+  def apply() = new CompilerHalt
+}
