+++ app/actors/messages/CompilerHalted.scala
@@ -1,0 +1,7 @@
+package actors.messages
+
+class CompilerHalted
+
+object CompilerHalted {
+  def apply() = new CompilerHalted
+}
