+++ app/actors/messages/CompilerOutput.scala
@@ -1,0 +1,6 @@
+package actors.messages
+
+import controllers.MathBotCompiler.ClientFrame
+import model.models.Problem
+
+case class CompilerOutput(frames : List[ClientFrame], problem: Problem)
