+++ app/actors/messages/CompilerStep.scala
@@ -1,0 +1,5 @@
+package actors.messages
+
+import model.models.Problem
+
+case class CompilerStep(steps : Int, problem : Problem)
