+++ app/actors/messages/SocketRequest.scala
@@ -1,0 +1,3 @@
+package actors.messages
+
+case class SocketRequest(steps : Option[Int], problem : Option[String], halt : Option[Boolean])
