+++ app/actors/messages/SomethingBroke.scala
@@ -1,0 +1,5 @@
+package actors.messages
+
+import play.api.libs.json.JsValue
+
+case class SomethingBroke(msg: JsValue)
